refactor: User 컨트롤러, 서비스 안에 있는 Favorite, Management 관련 코드 분리
