+++ server/module-api/src/test/java/com/momo/user/acceptance/step/UserAcceptanceStep.java
@@ -4,18 +4,11 @@ import static com.momo.CommonFileUploadSupport.uploadAssuredSupport;
 import static io.restassured.RestAssured.given;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.momo.domain.common.dto.EnumResponse;
-import com.momo.domain.group.dto.GroupCreateRequest;
-import com.momo.domain.user.dto.FavoriteCategoriesUpdateRequest;
-import com.momo.domain.user.dto.FavoriteGroupCardResponse;
-import com.momo.domain.user.dto.FavoriteGroupCreateRequest;
-import com.momo.domain.user.dto.ParticipatingGroupCardResponse;
 import com.momo.domain.user.dto.UserResponse;
 import com.momo.domain.user.dto.UserUpdateRequest;
 import com.momo.domain.user.entity.User;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import java.util.List;
 import org.junit.jupiter.api.Assertions;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
@@ -32,44 +25,6 @@ public class UserAcceptanceStep {
             () -> assertThat(response.getDistrict()).isEqualTo(user.getDistrict()),
             () -> assertThat(response.getUniversity()).isEqualTo(user.getUniversity()),
             () -> assertThat(response.getCategories().size()).isEqualTo(2)
-        );
-    }
-
-    public static void assertThatFindFavoriteGroups(List<FavoriteGroupCardResponse> responses,
-        GroupCreateRequest request) {
-        Assertions.assertAll(
-            () -> assertThat(responses.size()).isEqualTo(1),
-            () -> assertThat(responses.get(0).getId()).isNotNull(),
-            () -> assertThat(responses.get(0).getGroupCardResponse().getId()).isNotNull(),
-            () -> assertThat(responses.get(0).getGroupCardResponse().getName()).isEqualTo(request.getName()),
-            () -> assertThat(responses.get(0).getGroupCardResponse().getStartDate()).isEqualTo(request.getStartDate()),
-            () -> assertThat(responses.get(0).getGroupCardResponse().getImageUrl()).isNotNull(),
-            () -> assertThat(responses.get(0).getGroupCardResponse().getParticipantCnt()).isEqualTo(1)
-        );
-    }
-
-    public static void assertThatFindFavoriteCategories(List<EnumResponse> responses,
-        FavoriteCategoriesUpdateRequest request) {
-        Assertions.assertAll(
-            () -> assertThat(responses.size()).isEqualTo(request.getFavoriteCategories().size()),
-            () -> assertThat(responses.get(0).getName()).isEqualTo(request.getFavoriteCategories().get(0).getName()),
-            () -> assertThat(responses.get(0).getCode()).isEqualTo(request.getFavoriteCategories().get(0).getCode()),
-            () -> assertThat(responses.get(1).getName()).isEqualTo(request.getFavoriteCategories().get(1).getName()),
-            () -> assertThat(responses.get(1).getCode()).isEqualTo(request.getFavoriteCategories().get(1).getCode())
-        );
-    }
-
-    public static void assertThatFindParticipatingGroups(List<ParticipatingGroupCardResponse> responses,
-        GroupCreateRequest request) {
-        Assertions.assertAll(
-            () -> assertThat(responses.size()).isEqualTo(1),
-            () -> assertThat(responses.get(0).getId()).isNotNull(),
-            () -> assertThat(responses.get(0).getImageUrl()).isNotNull(),
-            () -> assertThat(responses.get(0).getName()).isEqualTo(request.getName()),
-            () -> assertThat(responses.get(0).getStartDate()).isEqualTo(request.getStartDate()),
-            () -> assertThat(responses.get(0).isOffline()).isEqualTo(request.getIsOffline()),
-            () -> assertThat(responses.get(0).isEnd()).isFalse(),
-            () -> assertThat(responses.get(0).getParticipantCnt()).isEqualTo(1)
         );
     }
 
@@ -95,68 +50,11 @@ public class UserAcceptanceStep {
         );
     }
 
-    public static ExtractableResponse<Response> requestToCreateFavoriteGroup(String token,
-        FavoriteGroupCreateRequest request) {
-        return given().log().all()
-            .header(HttpHeaders.AUTHORIZATION, "Bearer " + token)
-            .contentType(MediaType.APPLICATION_JSON_VALUE)
-            .body(request)
-            .when()
-            .post("/api/user/favorite-group")
-            .then().log().all()
-            .extract();
-    }
-
     public static ExtractableResponse<Response> requestToFindMyInformation(String token) {
         return given().log().all()
             .header(HttpHeaders.AUTHORIZATION, "Bearer " + token)
             .when()
             .get("/api/user")
-            .then().log().all()
-            .extract();
-    }
-
-    public static ExtractableResponse<Response> requestToFindFavoriteGroupCount(String token) {
-        return given().log().all()
-            .header(HttpHeaders.AUTHORIZATION, "Bearer " + token)
-            .when()
-            .get("/api/user/favorite-group-count")
-            .then().log().all()
-            .extract();
-    }
-
-    public static ExtractableResponse<Response> requestToFindFavoriteGroups(String token) {
-        return given().log().all()
-            .header(HttpHeaders.AUTHORIZATION, "Bearer " + token)
-            .when()
-            .get("/api/user/favorite-groups")
-            .then().log().all()
-            .extract();
-    }
-
-    public static ExtractableResponse<Response> requestToFindFavoriteCategories(String token) {
-        return given().log().all()
-            .header(HttpHeaders.AUTHORIZATION, "Bearer " + token)
-            .when()
-            .get("/api/user/favorite-categories")
-            .then().log().all()
-            .extract();
-    }
-
-    public static ExtractableResponse<Response> requestToFindParticipatingGroupCount(String token) {
-        return given().log().all()
-            .header(HttpHeaders.AUTHORIZATION, "Bearer " + token)
-            .when()
-            .get("/api/user/participating-group-count")
-            .then().log().all()
-            .extract();
-    }
-
-    public static ExtractableResponse<Response> requestToFindParticipatingGroups(String token) {
-        return given().log().all()
-            .header(HttpHeaders.AUTHORIZATION, "Bearer " + token)
-            .when()
-            .get("/api/user/participating-groups")
             .then().log().all()
             .extract();
     }
@@ -178,27 +76,6 @@ public class UserAcceptanceStep {
                 .header(HttpHeaders.AUTHORIZATION, "Bearer " + token)
                 .contentType(MediaType.MULTIPART_FORM_DATA_VALUE), userUpdateRequest)
             .post("/api/user/update")
-            .then().log().all()
-            .extract();
-    }
-
-    public static ExtractableResponse<Response> requestToUpdateFavoriteCategories(String token,
-        FavoriteCategoriesUpdateRequest request) {
-        return given().log().all()
-            .header(HttpHeaders.AUTHORIZATION, "Bearer " + token)
-            .contentType(MediaType.APPLICATION_JSON_VALUE)
-            .body(request)
-            .when()
-            .patch("/api/user/favorite-categories")
-            .then().log().all()
-            .extract();
-    }
-
-    public static ExtractableResponse<Response> requestToDeleteFavoriteGroup(String token, Long favoriteId) {
-        return given().log().all()
-            .header(HttpHeaders.AUTHORIZATION, "Bearer " + token)
-            .when()
-            .delete("/api/user/favorite-group/{id}", favoriteId)
             .then().log().all()
             .extract();
     }
