+++ server/module-api/src/test/java/com/momo/user/docs/FavoriteRestDocsTest.java
@@ -1,0 +1,128 @@
+package com.momo.user.docs;
+
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.when;
+import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.delete;
+import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.get;
+import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.patch;
+import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+import com.momo.RestDocsControllerTest;
+import com.momo.api.favorite.FavoriteController;
+import com.momo.domain.common.dto.EnumResponse;
+import com.momo.domain.group.dto.GroupCardResponse;
+import com.momo.domain.group.entity.Category;
+import com.momo.domain.favorite.dto.FavoriteCategoriesUpdateRequest;
+import com.momo.domain.favorite.dto.FavoriteGroupCardResponse;
+import com.momo.domain.favorite.dto.FavoriteGroupCountResponse;
+import com.momo.domain.favorite.dto.FavoriteGroupCreateRequest;
+import com.momo.domain.favorite.service.FavoriteService;
+import java.time.LocalDate;
+import java.util.List;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.mockito.InjectMocks;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.http.MediaType;
+
+@WebMvcTest(FavoriteController.class)
+@DisplayName("관심 문서화 테스트")
+public class FavoriteRestDocsTest extends RestDocsControllerTest {
+
+    @InjectMocks
+    private FavoriteController favoriteController;
+
+    @MockBean
+    private FavoriteService favoriteService;
+
+    @Test
+    void 관심_모임_등록() throws Exception {
+        FavoriteGroupCreateRequest request = FavoriteGroupCreateRequest.builder()
+            .groupId(1L)
+            .build();
+        String content = super.objectMapper.writeValueAsString(request);
+        super.mockMvc.perform(post("/api/favorite/group")
+                .content(content)
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+            )
+            .andDo(print())
+            .andExpect(status().isCreated())
+            .andDo(FavoriteDocumentation.createFavoriteGroup());
+    }
+
+    @Test
+    void 관심_모임_수_조회() throws Exception {
+        FavoriteGroupCountResponse response = new FavoriteGroupCountResponse(10L);
+        when(favoriteService.countFavoriteGroupsByUser(any())).thenReturn(response);
+        super.mockMvc.perform(get("/api/favorite/group/count"))
+            .andDo(print())
+            .andExpect(status().isOk())
+            .andDo(FavoriteDocumentation.findFavoriteGroupCount());
+    }
+
+    @Test
+    void 관심_모임_목록_조회() throws Exception {
+        List<FavoriteGroupCardResponse> responses = List.of(
+            FavoriteGroupCardResponse.builder()
+                .id(1L)
+                .groupCardResponse(
+                    GroupCardResponse.builder()
+                        .id(1L)
+                        .name("모임 이름")
+                        .imageUrl("이미지 URL")
+                        .startDate(LocalDate.of(2022, 1, 6))
+                        .isOffline(true)
+                        .participantCnt(5L)
+                        .isFavoriteGroup(true)
+                        .build()
+                )
+                .build()
+        );
+
+        when(favoriteService.findFavoriteGroupsByUser(any())).thenReturn(responses);
+
+        super.mockMvc.perform(get("/api/favorite/groups"))
+            .andDo(print())
+            .andExpect(status().isOk())
+            .andDo(FavoriteDocumentation.findFavoriteGroups());
+    }
+
+    @Test
+    void 관심_카테고리_목록_조회() throws Exception {
+        List<EnumResponse> responses = List.of(
+            EnumResponse.ofCategory(Category.LIFE),
+            EnumResponse.ofCategory(Category.HOBBY)
+        );
+
+        when(favoriteService.findFavoriteCategoriesByUser(any())).thenReturn(responses);
+
+        super.mockMvc.perform(get("/api/favorite/categories"))
+            .andDo(print())
+            .andExpect(status().isOk())
+            .andDo(FavoriteDocumentation.findFavoriteCategories());
+    }
+
+    @Test
+    void 관심_카테고리_수정() throws Exception {
+        FavoriteCategoriesUpdateRequest request = new FavoriteCategoriesUpdateRequest(List.of(Category.LIFE));
+        String content = super.objectMapper.writeValueAsString(request);
+        super.mockMvc.perform(patch("/api/favorite/categories")
+                .content(content)
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+            )
+            .andDo(print())
+            .andExpect(status().isOk())
+            .andDo(FavoriteDocumentation.updateFavoriteCategories());
+    }
+
+    @Test
+    void 관심_모임_삭제() throws Exception {
+        super.mockMvc.perform(delete("/api/favorite/group/{id}", 1L))
+            .andDo(print())
+            .andExpect(status().isNoContent())
+            .andDo(FavoriteDocumentation.deleteFavoriteGroup());
+    }
+}
