+++ server/module-api/src/test/java/com/momo/user/docs/UserDocumentation.java
@@ -18,15 +18,6 @@ import org.springframework.restdocs.request.RequestPartDescriptor;
 
 public class UserDocumentation {
 
-    public static RestDocumentationResultHandler createFavoriteGroup() {
-        FieldDescriptor[] request = new FieldDescriptor[]{
-            fieldWithPath("groupId").type(JsonFieldType.NUMBER).description("모임 ID")
-        };
-        return document("user/createFavoriteGroup",
-            requestFields(request)
-        );
-    }
-
     public static RestDocumentationResultHandler findMyInformation() {
         FieldDescriptor[] responseUser = new FieldDescriptor[]{
             fieldWithPath("id").type(JsonFieldType.NUMBER).description("유저 ID"),
@@ -50,70 +41,6 @@ public class UserDocumentation {
         };
         return document("user/validateDuplicateNickname",
             requestParameters(requestParam)
-        );
-    }
-
-    public static RestDocumentationResultHandler findFavoriteGroupCount() {
-        FieldDescriptor[] response = new FieldDescriptor[]{
-            fieldWithPath("count").type(JsonFieldType.NUMBER).description("관심 모임 수"),
-        };
-        return document("user/findFavoriteGroupCount",
-            responseFields(response)
-        );
-    }
-
-    public static RestDocumentationResultHandler findFavoriteGroups() {
-        FieldDescriptor[] response = new FieldDescriptor[]{
-            fieldWithPath("[].id").type(JsonFieldType.NUMBER).description("관심 모임 ID"),
-            fieldWithPath("[].groupCardResponse.id").type(JsonFieldType.NUMBER).description("모임 ID"),
-            fieldWithPath("[].groupCardResponse.name").type(JsonFieldType.STRING).description("모임 이름"),
-            fieldWithPath("[].groupCardResponse.imageUrl").type(JsonFieldType.STRING).description("모임 대표 이미지 URL"),
-            fieldWithPath("[].groupCardResponse.startDate").type(JsonFieldType.STRING).description("모임 시작 날짜"),
-            fieldWithPath("[].groupCardResponse.offline").type(JsonFieldType.BOOLEAN).description("모임 오프라인 여부"),
-            fieldWithPath("[].groupCardResponse.participantCnt").type(JsonFieldType.NUMBER).description("모임 참여자 수"),
-            fieldWithPath("[].groupCardResponse.favoriteGroup").type(JsonFieldType.BOOLEAN).description("관심 모임 여부")
-        };
-        return document("user/findFavoriteGroups",
-            responseFields(response)
-        );
-    }
-
-    public static RestDocumentationResultHandler findFavoriteCategories() {
-        FieldDescriptor[] response = new FieldDescriptor[]{
-            fieldWithPath("[].code").type(JsonFieldType.STRING).description("관심 카테고리 코드"),
-            fieldWithPath("[].name").type(JsonFieldType.STRING).description("관심 카테고리 이름")
-        };
-        return document("user/findFavoriteCategories",
-            responseFields(response)
-        );
-    }
-
-    public static RestDocumentationResultHandler findParticipatingGroupCount() {
-        FieldDescriptor[] response = new FieldDescriptor[]{
-            fieldWithPath("count").type(JsonFieldType.NUMBER).description("참여한 모임 수"),
-        };
-        return document("user/findParticipatingGroupCount",
-            responseFields(response)
-        );
-    }
-
-    public static RestDocumentationResultHandler findParticipatingGroups() {
-        FieldDescriptor[] response = new FieldDescriptor[]{
-            fieldWithPath("[].id").type(JsonFieldType.NUMBER).description("모임 ID"),
-            fieldWithPath("[].name").type(JsonFieldType.STRING).description("모임 이름"),
-            fieldWithPath("[].imageUrl").type(JsonFieldType.STRING)
-                .description("모임 대표 이미지 URL"),
-            fieldWithPath("[].startDate").type(JsonFieldType.STRING)
-                .description("모임 시작 날짜"),
-            fieldWithPath("[].offline").type(JsonFieldType.BOOLEAN)
-                .description("모임 오프라인 여부"),
-            fieldWithPath("[].end").type(JsonFieldType.BOOLEAN)
-                .description("모임 종료 여부"),
-            fieldWithPath("[].participantCnt").type(JsonFieldType.NUMBER)
-                .description("모임 참여자 수"),
-        };
-        return document("user/findParticipatingGroups",
-            responseFields(response)
         );
     }
 
@@ -142,23 +69,6 @@ public class UserDocumentation {
         return document("user/updateMyInformationWithImage",
             requestParameters(requestUser),
             requestParts(requestPart)
-        );
-    }
-
-    public static RestDocumentationResultHandler updateFavoriteCategories() {
-        FieldDescriptor[] requestUser = new FieldDescriptor[]{
-            fieldWithPath("favoriteCategories").type(JsonFieldType.ARRAY).description("관심 카테고리 목록"),
-        };
-        return document("user/updateFavoriteCategories",
-            requestFields(requestUser)
-        );
-    }
-
-    public static RestDocumentationResultHandler deleteFavoriteGroup() {
-        return document("user/deleteFavoriteGroup",
-            pathParameters(
-                parameterWithName("id").description("모임 ID")
-            )
         );
     }
 }
