+++ server/module-core/src/main/java/com/momo/domain/management/service/ManagementService.java
@@ -1,20 +1,25 @@
-package com.momo.domain.user.service;
+package com.momo.domain.management.service;
 
 import com.momo.domain.group.repository.ParticipantRepository;
-import com.momo.domain.user.dto.ParticipatingGroupCardResponse;
-import com.momo.domain.user.dto.ParticipatingGroupCountResponse;
+import com.momo.domain.post.entity.Post;
+import com.momo.domain.post.repository.PostRepository;
+import com.momo.domain.management.dto.MyPostCardResponse;
+import com.momo.domain.management.dto.ParticipatingGroupCardResponse;
+import com.momo.domain.management.dto.ParticipatingGroupCountResponse;
 import com.momo.domain.user.entity.User;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @Transactional
 @RequiredArgsConstructor
-public class GroupManagementService {
+public class ManagementService {
 
     private final ParticipantRepository participantRepository;
+    private final PostRepository postRepository;
 
     @Transactional(readOnly = true)
     public ParticipatingGroupCountResponse findParticipatingGroupCountByUser(User loginUser) {
@@ -25,5 +30,12 @@ public class GroupManagementService {
     @Transactional(readOnly = true)
     public List<ParticipatingGroupCardResponse> findParticipatingGroupsByUser(User loginUser) {
         return participantRepository.findParticipatingGroupsByUser(loginUser);
+    }
+
+    @Transactional(readOnly = true)
+    public List<MyPostCardResponse> findMyPostsByUser(User loginUser, int page, int size) {
+        List<Post> posts = postRepository
+            .findAllWithGroupAndAuthorByUserOrderByCreatedDateDesc(loginUser, PageRequest.of(page, size));
+        return MyPostCardResponse.listOf(posts);
     }
 }
