+++ server/module-core/src/test/java/com/momo/domain/management/service/ManagementServiceTest.java
@@ -1,4 +1,4 @@
-package com.momo.domain.user.service;
+package com.momo.domain.management.service;
 
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
 import static org.mockito.ArgumentMatchers.any;
@@ -7,7 +7,7 @@ import static org.mockito.Mockito.verify;
 
 import com.momo.common.ServiceTest;
 import com.momo.domain.group.repository.ParticipantRepository;
-import com.momo.domain.user.dto.ParticipatingGroupCountResponse;
+import com.momo.domain.management.dto.ParticipatingGroupCountResponse;
 import com.momo.domain.user.entity.User;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
@@ -15,14 +15,14 @@ import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 
-@DisplayName("모임 관리 서비스 테스트")
-public class GroupManagementServiceTest extends ServiceTest {
+@DisplayName("관리 서비스 테스트")
+public class ManagementServiceTest extends ServiceTest {
 
     @Mock
     private ParticipantRepository participantRepository;
 
     @InjectMocks
-    private GroupManagementService groupManagementService;
+    private ManagementService managementService;
 
     @Test
     void 유저가_참여한_모임_수를_조회한다() {
@@ -31,7 +31,7 @@ public class GroupManagementServiceTest extends ServiceTest {
 
         given(participantRepository.countAllByUser(any())).willReturn(expected);
 
-        ParticipatingGroupCountResponse actual = groupManagementService.findParticipatingGroupCountByUser(user);
+        ParticipatingGroupCountResponse actual = managementService.findParticipatingGroupCountByUser(user);
 
         verify(participantRepository).countAllByUser(any());
         Assertions.assertAll(
